Refactor: getDatesOfPeriod util 분리
