+++ src/utils/dateUtils.js
@@ -1,0 +1,13 @@
+const getDatesOfPeriod = (startDate, endDate) => {
+  const dates = [];
+  const currentDate = new Date(startDate);
+
+  while (currentDate <= endDate) {
+    dates.push(new Date(currentDate));
+    currentDate.setDate(currentDate.getDate() + 1);
+  }
+
+  return dates;
+};
+
+export default getDatesOfPeriod;
